test: separate out buffer slice tests
